Refactor: Extract io_readers from geom_io.py
